+++ apps/backend/src/app/api/latest/webhooks/stripe/route.tsx
@@ -1,10 +1,13 @@
+import { getProject } from "@/lib/projects";
 import { GLOBAL_STRIPE } from "@/lib/stripe";
 import { prismaClient } from "@/prisma-client";
 import { getEnvVariable } from "@stackframe/stack-shared/dist/utils/env";
+import { StackAssertionError } from "@stackframe/stack-shared/dist/utils/errors";
 import { NextApiRequest } from "next";
 import { headers } from "next/headers";
 import { Readable } from "node:stream";
 import Stripe from "stripe";
+import { STRIPE_CONNECT_EVENT_HANDLERS } from "./handler";
 
 // $ stripe listen --forward-to http://localhost:8102/api/v1/webhooks/stripe
 // $ stripe trigger account.updated
@@ -18,18 +21,6 @@ async function buffer(readable: Readable) {
 
 const stripe = GLOBAL_STRIPE;
 const STACK_STRIPE_WEBHOOK_SECRET = getEnvVariable("STACK_STRIPE_WEBHOOK_SECRET");
-
-type StripeConnectEventHandler<T extends Stripe.Event.Type> = (
-  stripe: Stripe,
-  event: Extract<Stripe.Event, { type: T }>,
-) => Promise<void>;
-
-const STRIPE_CONNECT_EVENT_HANDLERS: {
-  [T in Stripe.Event.Type]?: StripeConnectEventHandler<T>
-} = {
-  "customer.subscription.created": async (stripe, event) => {
-  },
-};
 
 export const POST = async (req: NextApiRequest) => {
   try {
@@ -49,8 +40,13 @@ export const POST = async (req: NextApiRequest) => {
       return Response.json({ error: `Webhook error: ${error}` }, { status: 400 });
     }
 
+    const project = await getProjectFromEvent(event);
+    if (!project) {
+      return Response.json({ error: 'Project not found' }, { status: 400 });
+    }
+
     // Handle the event
-    await STRIPE_CONNECT_EVENT_HANDLERS[event.type]?.(stripe, event as any);
+    await STRIPE_CONNECT_EVENT_HANDLERS[event.type]?.(stripe, event as any, project);
 
     return Response.json({ received: true });
   } catch (error) {
@@ -58,6 +54,26 @@ export const POST = async (req: NextApiRequest) => {
     return Response.json({ error: 'Webhook error' }, { status: 400 });
   }
 };
+
+
+async function getProjectFromEvent(event: Stripe.Event) {
+  if (event.account === undefined) {
+    // return the internal project
+    return await getProject("internal");
+  }
+
+  const paymentConfig = await prismaClient.paymentConfig.findFirst({
+    where: {
+      stripeAccountId: event.account,
+    },
+  });
+
+  if (!paymentConfig) {
+    throw new StackAssertionError("Payment config not found");
+  }
+
+  return await getProject(paymentConfig.projectConfigId);
+}
 
 export const config = {
   api: {
